Separate playlist tests
